指摘箇所修正しました

+++ database/migrations/2021_09_19_085341_create_planets_table.php
@@ -17,7 +17,7 @@ class CreatePlanetsTable extends Migration
             $table->increments('id');
             $table->string('name');
             $table->string('name_english');
-            $table->mediumInteger('radius');
+            $table->Integer('radius');
             $table->integer('weight');
             $table->timestamps();
         });
